+++ src/pararnn/cells/__init__.py
@@ -1,6 +1,7 @@
 from pararnn.cells.para_gru import ParaGRU
 from pararnn.cells.para_lstm import ParaLSTM
 from pararnn.cells.para_m2rnn import ParaM2RNN
+from pararnn.cells.para_nlru import ParaNLRU
 from pararnn.cells.para_slstm import ParaSLSTM
 from pararnn.cells.protocol import RNNCell, check_cell
 
@@ -8,6 +9,7 @@ __all__ = [
     "ParaGRU",
     "ParaLSTM",
     "ParaM2RNN",
+    "ParaNLRU",
     "ParaSLSTM",
     "RNNCell",
     "check_cell",